Move signOut and cookies to /refactor_2021.
