+++ packages/openneuro-app/src/scripts/refactor_2021/authentication/signOut.ts
