+++ packages/openneuro-app/src/scripts/refactor_2021/utils/cookies.js
