+++ packages/openneuro-app/src/scripts/test-utils.js
@@ -1,4 +1,4 @@
-import cookies from './utils/cookies.js'
+import cookies from './refactor_2021/utils/cookies.js'
 
 /**
  * Sets a login cookie during a test run
